[WIP] Create "clicker" example
Adds an oversimplified click counting web app. The sole functionality is
to increment a count on every post request. Clicking the button on the
page submits an empty post request.

I'd like to add some docs around setup, what an HTTP request is, etc
before getting this merged, so I've labeled it WIP for now.