Ryan's Demo
Just a little something I created for a data analysis demo. Wanted to
add a binary classifier section after the conclusion, but had
difficulties with matplotlib and plotting a decision boundary. I don’t
think it will take close to an hour to go through, but I worked hard on
it and didn’t want it to go to waste and not submit it.

The csv file contains the data, using the famous Titanic dataset. ipynb
file is what the user is met to run in a Jupyter notebook. It should be
able to run with both Python 2.7 and 3.6 (most of the code is from
modules). yml file contains all the dependancies (including Jupyter
stuff), allowing for an easy download of the environment used.

For myself, I used the Anaconda distribution which includes the package
manager Conda and Jupyter notebooks to run this on my Mac.

Take a look and see if it’s worthy. Use it if it’s good.